+++ quotebot/src/main/java/com/andmark/quotebot/service/command/ViewPendingCommand.java
@@ -1,0 +1,20 @@
+package com.andmark.quotebot.service.command;
+
+import lombok.extern.slf4j.Slf4j;
+import org.telegram.telegrambots.meta.api.objects.Chat;
+import org.telegram.telegrambots.meta.api.objects.User;
+import org.telegram.telegrambots.meta.bots.AbsSender;
+
+@Slf4j
+public class ViewPendingCommand extends QuoteCommand{
+
+    public ViewPendingCommand() {
+        super("view_pending", "View the dates of pending publications");
+    }
+
+    @Override
+    public void execute(AbsSender absSender, User user, Chat chat, String[] strings) {
+        log.debug("user with id = {} execute /view_pending", user.getId());
+
+    }
+}